+++ packages/apollo-tools/src/buildServiceDefinition.ts
@@ -13,7 +13,8 @@ import {
   isObjectType,
   SchemaDefinitionNode,
   OperationTypeNode,
-  SchemaExtensionNode
+  SchemaExtensionNode,
+  OperationTypeDefinitionNode
 } from "graphql";
 import { isNode, isDocumentNode } from "./utilities/graphql";
 import { GraphQLResolverMap } from "./schema/resolverMap";
@@ -27,6 +28,10 @@ export interface GraphQLSchemaModule {
 interface GraphQLServiceDefinition {
   schema?: GraphQLSchema;
   errors?: GraphQLError[];
+}
+
+function flattened<T>(arr: ReadonlyArray<ReadonlyArray<T>>): ReadonlyArray<T> {
+  return new Array<T>().concat(...arr);
 }
 
 export function buildServiceDefinition(
@@ -119,10 +124,11 @@ export function buildServiceDefinition(
     // but this matches the current 'last definition wins' behavior of `buildASTSchema`.
     const schemaDefinition = schemaDefinitions[schemaDefinitions.length - 1];
 
-    const operationTypes = [schemaDefinition, ...schemaExtensions]
-      .map(node => node.operationTypes)
-      .filter(isNotNullOrUndefined)
-      .flat();
+    const operationTypes = flattened(
+      [schemaDefinition, ...schemaExtensions]
+        .map(node => node.operationTypes)
+        .filter(isNotNullOrUndefined)
+    );
 
     for (const operationType of operationTypes) {
       const typeName = operationType.type.name.value;
@@ -178,15 +184,15 @@ export function buildServiceDefinition(
   }
 
   try {
-    const typeDefinitions = Object.values(typeDefinitionsMap).flat();
-    const directives = Object.values(directivesMap).flat();
+    const typeDefinitions = flattened(Object.values(typeDefinitionsMap));
+    const directives = flattened(Object.values(directivesMap));
 
     let schema = buildASTSchema({
       kind: Kind.DOCUMENT,
       definitions: [...typeDefinitions, ...directives]
     });
 
-    const typeExtensions = Object.values(typeExtensionsMap).flat();
+    const typeExtensions = flattened(Object.values(typeExtensionsMap));
 
     if (typeExtensions.length > 0) {
       schema = extendSchema(schema, {