+++ packages/apollo-tools/src/index.ts
@@ -1,5 +1,3 @@
-import "apollo-env";
-
 export * from "./utilities";
 
 export * from "./schema";