+++ CHANGELOG.md
@@ -25,8 +25,12 @@
 - `vscode-apollo`
   - <First `vscode-apollo` related entry goes here>
 
+## apollo-tools@0.5.1
+- Remove dependency on `apollo-env`, so using this package no longer installs polyfills.
+
 ## apollo-graphql@0.9.3
 - Complex directive arguments don't break `transformSchema` (Fixes #2162) [PR #2335](https://github.com/apollographql/apollo-tooling/pull/2335)
+
 ## apollo@2.33.2
 
 - Add a deprecation message to all `apollo service:*` commands, pointing people towards the [Apollo Rover CLI migration guide](https://go.apollo.dev/t/migration). <br/>